Instances are parsed once per run.sh call to optimize for running bench multiple times

+++ cleanup.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env bash
 
-mongo_process=$(pgrep mongod)
-kill ${mongo_process}
+# Kill MongoDB if running
+if pgrep mongod > /dev/null ; then
+   mongo_process=$(pgrep mongod)
+   kill ${mongo_process}
+fi
 
 rm results/log/mongodb.log
+rm bin/written_instances.json
